Move base-visualizer to new header dir pinocchio/visualizers
