+++ include/pinocchio/visualizers/base-visualizer.hpp
